Discord bot/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
coverage==7.2.5
cssselect==1.2.0
cymem==2.0.7
DateTime==5.1
discord-py-interactions==5.8.0
discord-typings==0.6.0
discord.py==2.3.1
diskcache==5.6.1
distlib==0.3.6
dnspython==2.3.0
docker==6.1.0
duckduckgo-search==2.9.3
emoji==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.12.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
googletrans==4.0.0rc1
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
identify==2.5.24
idna==2.10
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
langcodes==3.3.0
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.1.2
orjson==3.8.11
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.0
playsound==1.2.2
pluggy==1.0.0
pprintpp==0.4.0
pre-commit==3.3.1
preshed==3.0.8
protobuf==4.22.4
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
pygame==2.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.8.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.12.1
typer==0.7.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.23.0
wasabi==0.10.1
webdriver-manager==3.8.6
websocket-client==1.5.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0